Optimise text rendering in playback position display.
